Move logger config to CLI module
Previously we configured the root logger in the PMS package itself,
which would mean any code using the package as a library would get
the logging config as a side effect.

Rather than configure the root logger, which should be left to the
toplevel program, this re-scopes how we set the config to apply to
just the specific logger for the package.