Introduce Reading type to DRY-up reader internals
Reading DRYs-up knowledge of how to construct "RawData". Since Reader
classes don't know how to construct "ObsData", it makes sense that the
knowledge of how to construct "RawData" is also abstracted away.

The new type will also be heavily used in the following commits, as
we introduce a new interface to read from sensors. In future, the new
type could replace "RawData" and "ObsData" as a standard interface.