Separate reading single values from all values
This is part of adapting the reader module to support its use as a
library. Previously it was only possible to read *all* the data in
one go, which meant that error handling was hard to customise.

SensorReader.read_one uses StopIteration to indicate there are no
more readings. The choice of exception is to be consistent with the
MessageReader, but a custom one could potentially be used.