DRY-up Reader.__call__ using custom exceptions
This takes advantage of the new Stream classes to move towards a
single Reader class: SensorReader and MessageReader are now largely
the same except for the stream they use. Using a common definition
of Reader has the advantage of keeping their behaviour consistent.

Implementation notes:

- The sample counting code looks different between the two readers
but actually behaves the same way except that MessageReader would
only count once - at the class level.

- The "interval" default of "None" means MessageReader will continue
to behave as it did before, although in future it could of course
make use of this feature if needed.

- The two new "Stream" exceptions represent a layer of abstraction
that makes it possible for Reader.__call__ to make generic decisions
about Streams, rather than just sensors.