Simplify API to allow granular reader operation
In response to [^1].

Previously it was hard to operate the reader classes from calling
code that separates resource management ("open" / "close") and the
actual polling of data ("read_one"). This exposes new APIs to make
it easier to use reader classes in such a granular way.

Implementation notes:

- Reinstating a type hint for __enter__ of "-> Self" will become
possible in Python 3.11.

- "start" / "stop" were discussed in relation to the Sensor being
operated by a SensorReader, but these concepts don't make much sense
for the MessageReader class, so I've used "open" / "close" instead.

[^1]: https://github.com/avaldebe/PyPMS/pull/32#issuecomment-1324702383